backend/company/requirements.txt
requests==2.28.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
numpy==1.22.2
pandas==1.4.2